requirements.txt
helium==5.0.0
pandas==2.2.1
bs4==0.0.2
python-dotenv==1.0.1
